migrate all api calls from pages to use tanstack hooks
