+++ web/src/app/patients/page.tsx
@@ -1,10 +1,8 @@
 'use client'
 
-import { useState, useEffect } from 'react'
+import { useState } from 'react'
 import Link from 'next/link'
 import { useTranslations } from 'next-intl'
-import { patientsApi } from '@/lib/api'
-import { Patient } from '@/types'
 import { ProtectedRoute } from '@/components/auth/protected-route'
 import { AppLayout } from '@/components/layout/app-layout'
 import { Button } from '@/components/ui/button'
@@ -12,41 +10,15 @@ import { NewPatientForm } from '@/components/patients/new-patient-form'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import { Input } from '@/components/ui/input'
 import { Search, Plus } from 'lucide-react'
-import { useAuthStore } from '@/store/auth'
+import { useRecentPatients } from '@/hooks/use-patients'
 
 export default function PatientsPage() {
   const t = useTranslations('patients')
   const tCommon = useTranslations('common')
-  const [patients, setPatients] = useState<Patient[]>([])
-  const [isLoading, setIsLoading] = useState(true)
-  const [error, setError] = useState<string | null>(null)
   const [isNewPatientModalOpen, setIsNewPatientModalOpen] = useState(false)
 
-  const fetchPatients = async () => {
-    try {
-      setIsLoading(true)
-      setError(null)
-      // For now, just get recent patients - you can modify this to search
-      const patientData = await patientsApi.getRecent(50)
-      setPatients(patientData)
-    } catch (err) {
-      console.error('Error fetching patients:', err)
-      setError(err instanceof Error ? err.message : 'Error desconocido')
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  useEffect(() => {
-    // Only fetch patients if we have a clinic selected
-    const clinicId = useAuthStore.getState().clinicId
-    if (clinicId) {
-      fetchPatients()
-    } else {
-      setIsLoading(false)
-      setError('No clinic selected')
-    }
-  }, [])
+  // Use TanStack Query hook instead of manual state management
+  const { data: patients, isLoading, error } = useRecentPatients(50)
 
   if (isLoading) {
     return (
@@ -65,9 +37,9 @@ export default function PatientsPage() {
           <div className="flex items-center justify-center min-h-[400px]">
             <div className="text-center">
               <p className="text-destructive">
-                {error === 'No clinic selected'
+                {error.message === 'No clinic selected'
                   ? t('noClinicSelected')
-                  : `${t('loadingError')}: ${error}`}
+                  : `${t('loadingError')}: ${error.message}`}
               </p>
             </div>
           </div>
@@ -113,7 +85,7 @@ export default function PatientsPage() {
               </CardTitle>
             </CardHeader>
             <CardContent>
-              {patients?.length > 0 ? (
+              {patients && patients.length > 0 ? (
                 <div className="space-y-1">
                   {/* Header */}
                   <div className="grid grid-cols-4 gap-4 px-4 py-2 border-b text-sm font-medium text-muted-foreground">
@@ -162,7 +134,6 @@ export default function PatientsPage() {
           <NewPatientForm
             open={isNewPatientModalOpen}
             onOpenChange={setIsNewPatientModalOpen}
-            onSuccess={fetchPatients}
           />
         </div>
       </AppLayout>
